[AIR-3023] qs: eliminate redudant source code files
