+++ .augment/rules/ar-github-integration.md
@@ -2,6 +2,6 @@
 type: manual
 
 - Use GitHub CLI (gh) for all GitHub API operations
-- Implement retry mechanisms with helper.Retry() for network calls
+- Implement retry mechanisms with utils.Retry() for network calls
 - Include proper error messages for GitHub authentication failures
 - Follow the existing patterns in gitcmds/ package
