+++ internal/commands/consts.go
@@ -6,21 +6,8 @@ const (
 	msgOkSeeYou = "Ok, see you"
 	pushYes     = "y"
 
-	devDelParamFull = "delete"
-
-	devConfirm = "Dev branch '$reponame' will be created. Continue(y/n)? "
-
-	confMsgModFiles1 = "You have modified files: "
-	confMsgModFiles2 = "All will be kept not commted. Continue(y/n)?"
-
-	trueStr                 = "true"
-	oneSpace                = " "
 	EnvSkipQsVersionCheck   = "QS_SKIP_QS_VERSION_CHECK"
 	minimumCommitMessageLen = 8
-	fetch                   = "fetch"
-	origin                  = "origin"
-	git                     = "git"
-	refsNotes               = "refs/notes/*:refs/notes/*"
 )
 
 const (
