--- internal/commands/d.go
@@ -1,9 +1,0 @@
-package commands
-
-import (
-	"github.com/untillpro/qs/gitcmds"
-)
-
-func D(wd string) error {
-	return gitcmds.Download(wd)
-}
