--- internal/commands/g.go
@@ -1,7 +1,0 @@
-package commands
-
-import "github.com/untillpro/qs/gitcmds"
-
-func G(wd string) error {
-	return gitcmds.Gui(wd)
-}
