+++ internal/context/context.go
@@ -1,15 +1,3 @@
 package context
 
-type CtxKey int
 
-const (
-	CtxKeyCloneRepoPath CtxKey = iota
-	CtxKeyAnotherCloneRepoPath
-	CtxKeyDevBranchName
-	CtxKeyCustomBranchName
-	CtxKeyCreatedGithubIssueURL
-	CtxKeyJiraTicket
-	CtxKeyBranchPrefix
-	CtxKeyClipboard
-	CtxKeyCommitMessage
-)
