--- internal/helper/consts.go
@@ -1,8 +1,0 @@
-package helper
-
-const (
-	pushYes  = "y"
-	oneSpace = " "
-	maxDevBranchName = 50
-	msymbol          = "-"
-)
