+++ utils/branch_test.go
@@ -1,4 +1,4 @@
-package helper
+package utils
 
 import (
 	"testing"
